Safer escapes in removeFromHosts
On VMware, the machine ID is not a GUID like with VirtualBox, but the full path to the `.vmx` file (e.g. `/Users/dzuelke/.../blah.vmx`).

This fixes the `sed` call that removes entries from `/etc/hosts` by using a colon as the delimiter, and by escaping that delimiter (if that's not done, then it'll interpret ":" as the command due to the leading slash from the path).