Windows cmd scripts to change hosts permissions
Two windows cmd scripts that can work together to change permissions for
the current user to modify the hosts file.

win-hosts-check.cmd check if the permissions on the file are sufficient.

If not it will call win-hosts-set.cmd which will trigger the windows
permission dialog asking for enough rights to midify the permissions on
the hosts file for the current user.

See issue #40