Do not add entries that already exist
Modification of the system hostfile requires administrative privileges,
but some users may not want to grant those rights to this plugin. The
plugin already scans the hosts file for previously-existing entries, but
the heuristic is very strict and only matches machine-generated entries
as produced by the plugin itself.

Allow consumers to opt out of automatic hostfile modification (thereby
avoiding the need to grant additional privileges) by relaxing the
requirements for "matching" entries. Update the status messages to
document this behavior.