get hosts file path from VAGRANT_HOSTSUPDATER_PATH
This allows to set an environment variable VAGRANT_HOSTSUPDATER_PATH to be used to point to the file used as hosts file. 
This is nice when you do not want to change the /etc/hosts file or have some custom dns setup.
E.g. [dnsmasq](http://www.thekelleys.org.uk/dnsmasq/docs/dnsmasq-man.html) allows you to use any file or all files from a directory with the `--addn-hosts` option.