Fix intent to run each plugin once per host
We need to include both the class name and the machine ID to allow each action to be called once. (as a single run may call multiple actions)

Also the first call to `initialize()` seems to leave @machine.id empty which leads to incorrect hash being placed in /etc/hosts
So we ensure @machine.id has been provided before counting the call as the single run attempt.